Still issues with redirects and such but the reset feature now works and flows properly
