+++ src/main/java/org/launchcode/closettracker/AuthenticationFilter.java
@@ -1,7 +1,6 @@
 package org.launchcode.closettracker;
 
 import org.launchcode.closettracker.controllers.HomeController;
-import org.launchcode.closettracker.controllers.SessionController;
 import org.launchcode.closettracker.models.User;
 import org.launchcode.closettracker.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,9 +21,8 @@ public class AuthenticationFilter implements HandlerInterceptor {
     @Autowired
     HomeController homeController;
 
-    SessionController sessionController;
-
-    private static final List<String> whitelist = Arrays.asList("/index", "create", "/user/reset", "/user/reset-int", "/user/update", "/css");
+    private static final List<String> whitelist = Arrays.asList("/index", "/create", "/user/reset", "/styles.css",
+            "/user/reset-int", "/user/update", "/css");
 
     private static boolean isWhitelisted(String path) {
         for (String pathRoot : whitelist) {
@@ -43,20 +41,26 @@ public class AuthenticationFilter implements HandlerInterceptor {
         // Don't require sign-in for whitelisted pages
         if (isWhitelisted(request.getRequestURI())) {
             // returning true indicates that the request may proceed
+            String requestStuff = request.getRequestURI();
+            boolean checkIndexPath = requestStuff.equals("/index");
             return true;
         }
+        else {
+            HttpSession session = request.getSession();
+            try {
+                User user = homeController.getUserFromSession(session);
 
-        HttpSession session = request.getSession();
-        User user = sessionController.getUserFromSession(session);
+                // The user is logged in
+                if (user != null) {
+                    return true;
+                }
 
-        // The user is logged in
-        if (user != null) {
-            return true;
+                // The user is NOT logged in
+                response.sendRedirect("/index");
+                return false;
+            } catch (Exception exception) {
+                return false;
+            }
         }
-
-        // The user is NOT logged in
-        response.sendRedirect("/index");
-        return false;
     }
-
 }
