+++ src/main/java/org/launchcode/closettracker/controllers/HomeController.java
@@ -2,7 +2,6 @@ package org.launchcode.closettracker.controllers;
 
 import org.launchcode.closettracker.models.User;
 import org.launchcode.closettracker.models.dto.UpdatePasswordDTO;
-import org.launchcode.closettracker.models.dto.UserDTO;
 import org.launchcode.closettracker.models.dto.LoginFormDTO;
 import org.launchcode.closettracker.repositories.ItemRepository;
 import org.launchcode.closettracker.repositories.UserRepository;
@@ -17,9 +16,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
-import static org.launchcode.closettracker.controllers.UserController.goUserUpdate;
+import java.util.Optional;
 
 @Controller
 public class HomeController {
@@ -30,11 +30,29 @@ public class HomeController {
     @Autowired
     private ItemRepository itemRepository;
 
-    SessionController sessionController;
+    public static final String userSessionKey = "user";
 
-// Thymeleaf template page strings
+// Thymeleaf global page template strings
     public static final String goIndex = "index";
+    public static final String goRedirect = "redirect:";
+    public static final String goRedirectIndex = "redirect:/index";
     private static final String goRedirectUserUpdate = "redirect:user/update";
+    private static final String goRedirectUserCloset = "redirect:items/closet";
+
+// Function to retrieve userid from browser session
+    public User getUserFromSession(HttpSession session) {
+        Optional<User> user = userRepository.findById((Integer) session.getAttribute(userSessionKey));
+        if (user.isPresent()) {
+            return user.get();
+        }
+        else {
+            return null;
+        }
+    }
+
+    private void setUserInSession(HttpSession session, User user) {
+        session.setAttribute(userSessionKey, user.getId());
+    }
 
     public String home(HttpServletResponse response) {
         //create a cookie with name 'website' and value 'javapointers'
@@ -48,7 +66,7 @@ public class HomeController {
         return "home";
     }
 
-    //localhost:8080  Shows login form
+// User --> Show login form
     @GetMapping("/index")
     public String index (Model model){
         model.addAttribute("title", "Welcome to Closet Tracker");
@@ -56,57 +74,48 @@ public class HomeController {
         return goIndex;
     }
 
+    // User --> Process login form
     @PostMapping("/index")
     public String processLoginForm(@ModelAttribute @Valid LoginFormDTO loginFormDTO, Errors errors,
                                    HttpServletRequest request, Model model){
         if (errors.hasErrors()) {
             model.addAttribute("title", "Welcome to Closet Tracker");
-            model.addAttribute("errorMsg", "Entry not valid!");
             return goIndex;
         }
+
         User theUser = userRepository.findByEmail(loginFormDTO.getEmail());
 
-        // Is the user has reset their password, this checks to see if the flag is true.
-        // If true, the user is redirected to the update page to choose a new password
-
+    // Is the user has reset their password, this checks to see if the flag is true.
+    // If true, the user is redirected to the update page to choose a new password
         if (theUser.isPasswordReset()) {
-            model.addAttribute(new UpdatePasswordDTO());
-// May need to use this line instead of showing the update page as there is some issue with it showing but not working upon submit
-//            errors.rejectValue("title", "password.reset", "The password for this account was reset so you must create a new password before logging in.");
-
             model.addAttribute(new UpdatePasswordDTO());
             return goRedirectUserUpdate;
         }
 
+    // A final check to ensure there is
         if (theUser == null) {
             errors.rejectValue("email", "user.invalid", "Not a valid user");
             model.addAttribute("title", "Welcome to Closet Tracker");
             return goIndex;
         }
+
         String password = loginFormDTO.getPassword();
 
+    // I don't believe this works correctly; still allows log in with an incorrect password
         if (!theUser.isEncodedPasswordEqualsInputPassword(password)) {
             errors.rejectValue("password", "password.invalid", "Invalid password");
             model.addAttribute("title", "Welcome to Closet Tracker");
             return goIndex;
         }
-        sessionController.setUserInSession(request.getSession(), theUser);
+        setUserInSession(request.getSession(), theUser);
 
-        return "redirect:items/";
-    }
-
-    //localhost:8080/create
-    @GetMapping("create")
-    public String displayCreateAccountForm(Model model) {
-        model.addAttribute(new UserDTO());
-        model.addAttribute("title", "Create User Account");
-        return "create";
+        return goRedirectUserCloset;
     }
 
     @GetMapping("/logout")
     public String logout(HttpServletRequest request){
         request.getSession().invalidate();
-        return "redirect:";
+        return goRedirectIndex;
     }
